Moves several aliases to their right path
