--- src/main/scala/package.scala
@@ -1,17 +1,0 @@
-import scalaz._, Scalaz._
-
-package object `statelesser` {
-
-  type TypeNme = String
-  type OpticNme = String
-
-  type Symbol = String
-
-  def fresh: State[Stream[String], String] =
-    for {
-      s <- get[Stream[String]]
-      _ <- modify[Stream[String]](_.tail)
-    } yield s.head
-
-}
-
